+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Development
 
+* Added support to delete container tree hierarchies
 * Add ODA trigger that inserts a dlm.data_item when an execution block is added.
 * Added dlm-archive stfp endpoint. 
 * Added pg_sphere to local db build.